pyproject.toml
[tool.poetry]
name = "guardrail-comparison"
version = "0.1.0"
description = "trialling leading frameworks for LLM guardrails"
authors = ["m-misiura <mmisiura@redhat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
nemoguardrails = "^0.9.1.1"
guardrails-ai = "^0.5.4"
transformers = "^4.44.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
ipykernel = "^6.29.5"
black = "^24.8.0"
notebook = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
